+++ lib/get_state.dart
@@ -1,4 +1,5 @@
 library getstate;
 
+export 'package:get_it/get_it.dart';
 export 'package:get_state/src/abs.dart';
 export 'package:get_state/src/recorder.dart';