+++ lib/src/vm/base_view_model.dart
@@ -5,12 +5,12 @@
 part of '../abs.dart';
 
 ///
-/// 推荐配合一个或多个 [LiveData] / [LiveModel] 使用
+/// 适用于一个或多个 [LiveData] / [LiveModel] 使用
 /// 数据刷新逻辑都转移到 [LiveData] / [LiveModel] 中
-/// todo 后期实现与LiveData绑定, 当没有ViewModel没有被任何View使用时,VM应当让LiveData感知到
 abstract class BaseViewModel extends ChangeNotifier {
-
   bool get hasListeners => super.hasListeners;
+
+  void get vmNotify => super.notifyListeners();
 
   /// 相当于 State<>类中的 dispose();方法
   @protected